Isolate test files
The test suite had the following issues:

* A failing test would leave files in the `tests` directory, causing
  other tests to fail unexpectedly.
* Even successful tests would require clean up code like `sh.rm(...)` to
  avoid leaving test files.  This was easy to forget as shown by #153.
* The same test file would be used for several tests, making them unsafe
  to run in parallel.

This commit fixes that by using the `dotenv_file` and `tmp_path`
fixtures whenever the filesystem is required.  The advantage of those
fixtures is that they use unique test directories, which are eventually
cleaned up.

The `tmp_path` fixture is preferred to the `tmpdir` fixture because
`tmpdir` relies on `py.path`, a library in "maintenance mode".