Refactor parser to fix inconsistencies (#180)
* Move parser to separate module

* Add tests

* Use unicode strings for unit tests in Python 2

Using `str` (e.g. `bytes`) is inconsistent with the types and the
implementation.

* Refactor parser

This fixes inconsistencies reported after the release of version 0.10.0:

* Valid escapes were interpreted as control characters even when in
  single-quoted strings.
* `#` was interpreted as the start of a comment even if there was no
  whitespace preceding it.

However, we are keeping the interpretation of escapes in double-quoted
strings as they didn't make sense in versions before 0.10.0.

The single large regular expression is replaced with a handwritten
top-down parser using smaller regular expressions.  The reason for this
change is that it would have been very difficult or impossible to
satisfy the parsing requirements with a single regex.