Fix handling of lines with just a comment
A .env file like the following would trigger a warning:

    # comment

The problem is that with the previous `Binding` structure, it was
impossible to distinguish the parsing of such a line and that of a
malformed line.  Therefore, this commit adds an `error` field so that
error reporting works correctly in this case.