Re-simplify load_dotenv check for dotenv source
theskumar/python-dotenv#164
theskumar/python-dotenv#245
theskumar/python-dotenv#263
theskumar/python-dotenv@33cf7f4

The syntax will now be closer to the original. The only change to f is
the addition of find_dotenv() to check the dotenv_path. If find_dotenv
can't find a dotenv source at dotenv_path, the conditional expression
will move to the next or.