Improve quoting of values in `set_key`
The value of `quote_mode` must now be one of `auto`, `never` or
`always`, to ensure that users aren't accidentally relying on any other
value for their scripts to work.

Surrounding quotes are no longer stripped.  This makes it possible for
the user to control exactly what goes in the .env file.  Note that when
doing `dotenv set foo 'bar'` in Bash, the shell will have already
removed the quotes.

Single quotes are used instead of double quotes.  This avoids
accidentally having values interpreted by the parser or Bash (e.g. if
you set a password with `dotenv set password 'af$rb0'`.

Previously, the `auto` mode of quoting had the same effect as `always`.
This commit restores the functionality of `auto` by not quoting
alphanumeric values (which don't need quotes).  Plenty of other kinds of
values also don't need quotes but it's hard to know which ones without
actually parsing them, so we just omit quotes for alphanumeric values,
at least for now.