Require Python >= 3.5
This is a big change.  It will make it possible to simplify the code,
add more features, improve the robustness and lower the barrier to new
contributions.

As per [Python's packaging documentation][doc], the `python_requires`
keyword argument needs `setuptools >= 24.2.0` (released in 2016) and
will only have en effect for `pip >= 9.0.0` (released in 2016 as well).

[doc]: https://packaging.python.org/guides/distributing-packages-using-setuptools/#python-requires