Support concatenated strings
With an option to have a different behavior for single quoted
strings it would be nice to use mixed string (e.g. "ffo"'${bar}'
to produce "foo${bar}".

Intruduced it by reading multiple strings on the value side.
Unquoted strings are tricky because of the whitespace and comments.
All the rest should come for free, as it is a change from
handling a single value to handling a list of values and concatenation.